refactor(ensnode-sdk): move `ENSNamespaceIds` and `PluginName`
Put the types into relevant SDK modules.
